refactor(storage): move SwiftData models into active schema
Adopt a schema-owned KMReaderSchemaV6 model graph and expose the active models through top-level typealiases so runtime fetch and insert types match the active schema.

Keep model behavior in extensions colocated with the typealiases, and retain an inferred lightweight fallback for legacy runtime-backed V4/V5 stores that staged migration cannot identify.
