+++ KMReader/Features/Browse/Models/SavedFilter.swift
@@ -32,18 +32,9 @@ enum SavedFilterType: String, CaseIterable, Identifiable {
   }
 }
 
-typealias SavedFilter = SavedFilterV1
+typealias SavedFilter = KMReaderSchemaV6.SavedFilterV1
 
-@Model
-final class SavedFilterV1 {
-  @Attribute(.unique) var id: UUID
-
-  var name: String
-  var filterTypeRaw: String
-  var filterDataJSON: String
-  var createdAt: Date
-  var updatedAt: Date
-
+extension SavedFilter {
   var filterType: SavedFilterType {
     get {
       SavedFilterType(rawValue: filterTypeRaw) ?? .series
@@ -51,22 +42,6 @@ final class SavedFilterV1 {
     set {
       filterTypeRaw = newValue.rawValue
     }
-  }
-
-  init(
-    id: UUID = UUID(),
-    name: String,
-    filterType: SavedFilterType,
-    filterDataJSON: String,
-    createdAt: Date = Date(),
-    updatedAt: Date = Date()
-  ) {
-    self.id = id
-    self.name = name
-    self.filterTypeRaw = filterType.rawValue
-    self.filterDataJSON = filterDataJSON
-    self.createdAt = createdAt
-    self.updatedAt = updatedAt
   }
 
   @MainActor
